+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2023-02-19
+### Changed
+- `parseNote()` now accepts the note name and the note content instead of the path to the file.
+Browsers can be supported this way
+
+### Added
+- It's now possible use it in the Browser (couldn't add tests though)
+
+## [0.0.5] - 2023-02-19
+### Changed
+- Update dependencies
+- Update assets list
+
 ## [0.0.4] - 2022-12-30
 ### Fixed
 - Fix `CHANGELOG.md` typos